+++ docs/docs/change_log.rst
@@ -4,6 +4,7 @@ Change Log
 Current (0.1.14.dev)
 
 - Added the ability for partial classes to subclass from another partial struct or ``ctypes.Structure``.
+- Change mod folder loading so that it will load from sub-folders as well. (`#64 <https://github.com/monkeyman192/pyMHF/issues/64>`_)
 
 0.1.13 (26/06/2025)
 -------------------